Now set difficulty level of animal & everyday noises game the same as other modes
Previously, in the animal & everyday noises games, users pressed A to
play the game and B to learn the animals/noises. In every other mode,
users press dot 1 and dot 2 (and sometimes dot 3) to choose between
submodes. This patch updates the game logic and sound files so that
animal & everyday noises now work like all the other modes.